wrote data_io_test.jl, just testing the return type of data file.

+++ src/data_io.jl
@@ -4,7 +4,7 @@ function get_geno_data(file)
     # G_prob = convert(Array{Float32,2},readdlm("../data/hippocampus-genopr-AA-BB.csv", ','; skipstart=1)[:,2:end])
     # G = G_prob[:, 1:2:end]
 
-    geno_prob = convert(Array{Float32,2},readdlm(file, ','; skipstart=1)[:,2:end])
+    geno_prob = convert(Array{Float64,2},readdlm(file, ','; skipstart=1)[:,2:end])
     return geno_prob[:,1:2:end]
 end
 
@@ -14,8 +14,7 @@ function get_pheno_data(file)
     # Y = convert(Array{Float32,2}, pheno[:, 1:end])
 
     pheno = readdlm(file, ','; skipstart=1)[:,2:end-1]
-    return convert(Array{Float32,2}, pheno[:, 1:end])
-
+    return convert(Array{Float64,2}, pheno[:, 1:end])
 end
 
 # function extract_gmap()
